refactor: Reorder polygon rendering in game component
The rendering of the polygon element in the game component has been updated. It is now rendered after the main game elements, alleviating an issue where it was overlaying other elements. CSS code for the polygon was also moved to a different section in the game.module.css file for better code organization.