fix: Update color property in footer CSS components
The update targets color properties of .twitter, .facebook, and .github selectors in the footer.module.css file. The color property value of .twitter and .facebook is now changed only upon focus or hover, while the color of .github has been changed. This results in a more interactive and dynamic change of color properties on user interaction.