refactor: Refactor header navigation links and CSS styles
In this commit, the header navigation component has been refactored to use the new Link module. As part of this, the `.activated` CSS class and its related attributes in the navigation component were removed, as the new Link component comes with an 'isDisabled' prop, eliminating the need for additional CSS to control pointer events. This change enhances flexibility and consistency for link styling and behaviour across the application.