refactor: Remove custom Link component and refactor usage
Eliminated the custom Link component from the project and used the Link component available in the library directly in the header and navigation components. This refinement simplifies the codebase and aligns it better with the library's implementation standards, thereby improving maintainability. Also, the 'aria-current' attribute has been correctly implemented to mark the active link.

Signed-off-by: taichi221228 <taichi221228@icloud.com>