refactor: Refactor CSS classes in Header and Navigation components
The `.link` CSS class in Header and Navigation components was simplified. The class was removed from JSX file and replaced with tag identity in CSS, improving the readability of the code. This change reflects an efficient usage of CSS selectors, making it easier to identify elements and maintain the code in future.

Signed-off-by: taichi221228 <taichi221228@icloud.com>