refactor: Refactor border style for navigation list items in CSS
In the 'navigation.module.css' file, the border style for list items has been refined. The 'border-inline-end' has been removed from every list item and is now applied only to those that are not the last child in the list. Consequently, the explicit absence of an end border for the last child has also been removed.

Signed-off-by: taichi221228 <taichi221228@icloud.com>