refactor: Remove activated class from sidebar buttons
The "activated" class was removed from the sidebar buttons in the about-me route as it was not necessary. The changes reflect the removal of these class calls in the corresponding code blocks. Additionally, the CSS was updated to maintain consistent styling for these buttons when they are in disabled, focus-visible, or hover states.

Signed-off-by: taichi221228 <taichi221228@icloud.com>