refactor: Remove class dynamic assignment in activitybar
The dynamic assignment of the class property in the 'activitybar' component was removed since it was unnecessary. Additionally, in the accompanying CSS file, the selector has been changed from '.activated' to '[disabled]', modifying the conditions under which styles will be applied to enhance usability and compatibility across different states of the button elements.

Signed-off-by: taichi221228 <taichi221228@icloud.com>