feat: Update Tab component and close functionality
The changes primarily address updates to the Tab component, including the addition of a close functionality. The tab items now accept an object containing name, onClick$, and onClose$ properties. The style updates in "tab.module.css" contribute to the visual display of these changes. The updated functionality enables the close feature on each tab for improved user experience.

Signed-off-by: taichi221228 <taichi221228@icloud.com>