refactor: Remove 'KeyOfMap' type and replace 'Map' with 'Record'
The 'KeyOfMap' type in 'map.d.ts' has been removed due to its redundancy. With this change, any usage of 'Map' has been replaced with 'Record', which simplifies the handling of key-value pairs. This is seen in instances such as 'activities' and 'sides' in the 'about-me/index.tsx' file, and the 'activities' usage in 'activitybar.tsx'. This change enhances the readability of the code and simplifies the process of managing these objects.

Signed-off-by: taichi221228 <taichi221228@icloud.com>