refactor: Refactor activities data structure in about-me route
The 'activities' data structure in 'about-me' route has been converted from an object to an array for better management. Objects with keys representing each activity's name have been replaced with an array of objects, each containing name and Icon properties. This ensures that data is more consistently structured, improving the readability and maintainability of the code.

Signed-off-by: taichi221228 <taichi221228@icloud.com>