refactor: Update color definitions and blending techniques
This commit updates the color definitions from RGB to CSS color-mix in the application's CSS files. Removed RGB color variables in the global.css file and updated other component CSS files to use the color-mix for color blending. This approach simplifies the management of colors and is more compatible with different color spaces.

Signed-off-by: taichi221228 <taichi221228@icloud.com>