feat: Update Accordion to adjust to container's height
Additional features have been added to the Accordion component that allow it to resize according to the height of the container it's wrapped around. A ResizeObserver is used to track the offsetHeight of the container, which is then used to adjust the maxBlocksize of the Accordion. A new reference to the container DIV and additional signal elements were imported to facilitate these changes.

Signed-off-by: taichi221228 <taichi221228@icloud.com>