chore: Update component scaffold configuration
The configuration for the component scaffold in .scaffdog/component.md has been updated. The 'package' option has been removed from the 'choices' section and replaced with a 'confirm' option named 'isPackage', allowing for more accurate identification of whether a component is a package or not. The checks in the scaffolding code have been updated accordingly to reflect this configuration change.