chore: Remove 'isPage' functionality from scaffdog component
The 'isPage' feature has been removed from the scaffdog component. Previously, the 'isPage' question was asked and used to determine whether the component is for routing. Now, whether the component is for routing will be determined solely by the 'inRoutes' question. Additionally, the export statement has been simplified to always use the constant format.