feat: Update `<LinkAccordion />` and styles
The LinkAccordion component now leverages the Accordion component and allows for custom prop configurations. This includes defining 'head' and 'shouldOpen' properties. Additionally, we've updated the link-accordion.module.css to define a .body class that includes specifications for list items. This refactoring not only enhances performance but also improves maintainability of the code.