feat: Move `padding-block` from `<Body />` to `<Form />`
A new CSS module is added to style the contact form. The previous padding properties in the body's CSS are moved to a new container class in the form module, to ensure consistent styling. This change also includes the importation and usage of the module styles in the form's TypeScript file.